Casting np.round() result into float

+++ qiskit_experiments/library/characterization/rabi.py
@@ -161,7 +161,9 @@ class Rabi(BaseExperiment, RestlessMixin):
         # Create the circuits to run
         circs = []
         for amp in self.experiment_options.amplitudes:
-            amp = np.round(amp, decimals=6)
+            # casting is needed because for amplitude '0', np.round method return datatype of int32
+            # which isn't serializable in the metadata.
+            amp = float(np.round(amp, decimals=6))
             assigned_circ = circuit.assign_parameters({param: amp}, inplace=False)
             assigned_circ.metadata = {
                 "experiment_type": self._type,
